// src/types.ts
export type Project = {
  title: string;
  techStack: string[];
  description: string;
  image: string;
  github?: string;
  demo?: string;
  froentEnd?: string;
  backEnd?: string;
};

export type Experience = {
  companyName: string;
  image: string;
  shortDescription: string;
  date: string;
  longDescription: string [];
} 